# swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.3.0",
    "title": "Majority Judgment Server",
    "description": "MJS API documentation"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    },
    {
      "url": "https://localhost:3000/api"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/poll/create": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Newly created poll ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": "abcde1234",
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "string",
                  "example": "abcde1234",
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Auth_user": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "caca"
                  },
                  "choices": {
                    "type": "Array of strings",
                    "example": [
                      "Candidate A",
                      "Candidate B",
                      "..."
                    ]
                  },
                  "grades": {
                    "optional": true,
                    "type": "Array of strings",
                    "example": [
                      "A",
                      "B",
                      "C",
                      "D",
                      "E",
                      "F"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/poll/{id}/start": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Poll start date (JS date)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "number",
                  "example": 1647896248468,
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "number",
                  "example": 1647896248468,
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Auth_user": []
          }
        ]
      }
    },
    "/poll/{id}/close": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Poll end date (JS date)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "number",
                  "example": 1647896248468,
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "number",
                  "example": 1647896248468,
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Auth_user": []
          }
        ]
      }
    },
    "/poll/{id}/voter": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operation success flag",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean",
                  "example": true,
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "boolean",
                  "example": true,
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Auth_user": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "example": "voter_token_1234"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operation success flag",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean",
                  "example": true,
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "boolean",
                  "example": true,
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Auth_user": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "example": "voter_token_1234"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/poll/{id}/vote": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operation success flag",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean",
                  "example": true,
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "boolean",
                  "example": true,
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Auth_user": []
          }
        ]
      }
    },
    "/poll/{id}/results": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "Auth_user": {
        "type": "apiKey",
        "description": "Admin or Voter auth token",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": {
    "Auth_user": []
  }
}